requirements.txt
attrs==20.3.0
blis==0.2.4
boto3==1.14.55
botocore==1.17.55
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
cymem==2.0.4
Cython==0.29.21
doc2text==0.2.4
docutils==0.15.2
docx2txt==0.8
Flask==1.1.2
Flask-Cors==3.0.9
future==0.18.2
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
MarkupSafe==1.1.1
mime==0.1.0
murmurhash==1.0.4
nltk==3.5
numpy==1.19.1
pandas==1.1.1
pdfminer.six==20201018
pdfplumber==0.5.24
phonenumbers==8.12.12
Pillow==8.0.1
plac==0.9.6
preshed==2.0.1
pycparser==2.20
pycryptodome==3.9.9
PyPDF2==1.26.0
pyresparser==1.0.6
pyrsistent==0.17.3
pytesseract==0.3.6
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
regex==2020.10.28
requests==2.24.0
resume-parser==0.6
s3transfer==0.3.3
scipy==1.5.4
six==1.15.0
sortedcontainers==2.2.2
spacy==2.1.9
srsly==1.0.3
stemming==1.0.1
thinc==7.0.8
tika==1.24
tqdm==4.51.0
urllib3==1.25.10
Wand==0.6.3
wasabi==0.8.0
Werkzeug==1.0.1
zipp==3.4.0
